fix(deno-udd): Update Deno dependencies
Add `getSnapshot` and `getTimestampIds`

Remove `getSnapshots`, `readSnapshotsBulk` and `readSnapshots`
