+++ .github/workflows/udd.yml
@@ -16,7 +16,7 @@ jobs:
       - name: Update dependencies
         run: >
           deno run --allow-net --allow-read --allow-write=deps/
-          --allow-run=deno https://deno.land/x/udd@0.7.2/main.ts deps/*.ts
+          --allow-run=deno https://deno.land/x/udd@0.8.2/main.ts deps/*.ts
           --test="deno test --allow-read"
       - name: Create Pull Request
         uses: peter-evans/create-pull-request@v3
