+++ deps/scrapbox.ts
@@ -2,8 +2,8 @@ export type {
   BaseLine,
   Line,
   Scrapbox,
-} from "https://raw.githubusercontent.com/scrapbox-jp/types/0.7.1/userscript.ts";
+} from "https://raw.githubusercontent.com/scrapbox-jp/types/0.8.0/userscript.ts";
 export type {
   BaseStore,
-} from "https://raw.githubusercontent.com/scrapbox-jp/types/0.7.1/baseStore.ts";
+} from "https://raw.githubusercontent.com/scrapbox-jp/types/0.8.0/baseStore.ts";
 export * from "https://esm.sh/@progfay/scrapbox-parser@9.0.0";
